ci: Added acr_low_contrast_object_detectability to CI tests
